refactor: Extract amd import autofix to separate file
